third_party/blink/web_tests/fast/xsl/subframe-location-expected.txt

CONSOLE WARNING: crbug.com/435623334: This page uses XSLT, which being considered for removal from the web. If that happens, it is possible that this page will need to be updated to maintain functionality.
ALERT: SUCCESS
Test that a document in a subframe processed with XSLT has a correct location. (bug 18259).
